Add function to logout
